backend/requirements.txt
alembic==1.5.8
amqp==5.0.6
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
boto3==1.17.73
botocore==1.20.73
celery==5.0.5
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.0.0
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-mock==3.5.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
s3transfer==0.4.2
six==1.15.0
SQLAlchemy==1.4.4
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1